Step 13.3: Use GraphQL Modules
