+++ index.ts
@@ -1,71 +1,23 @@
 import { ApolloServer } from 'apollo-server-express';
-import cookie from 'cookie';
+import { GraphQLModule } from '@graphql-modules/core';
 import http from 'http';
-import jwt from 'jsonwebtoken';
 import { app } from './app';
-import { pool } from './db';
-import { origin, port, secret } from './env';
+import { origin, port } from './env';
 import { MyContext } from './context';
-import sql from 'sql-template-strings';
 import { UnsplashApi } from './schema/unsplash.api';
-const { PostgresPubSub } = require('graphql-postgres-subscriptions');
 
-import * as commonModule from './modules/common';
-import * as usersModule from './modules/users';
-import * as chatsModule from './modules/chats';
+import usersModule from './modules/users';
+import chatsModule from './modules/chats';
 
-const pubsub = new PostgresPubSub({
-  host: 'localhost',
-  port: process.env.DB_PORT ? parseInt(process.env.DB_PORT) : 5432,
-  user: 'testuser',
-  password: 'testpassword',
-  database: 'whatsapp',
+export const rootModule = new GraphQLModule({
+  name: 'root',
+  imports: [usersModule, chatsModule],
 });
+
 const server = new ApolloServer({
-  modules: [commonModule, usersModule, chatsModule],
-  context: async (session: any) => {
-    // Access the request object
-    let req = session.connection
-      ? session.connection.context.request
-      : session.req;
-
-    // It's subscription
-    if (session.connection) {
-      req.cookies = cookie.parse(req.headers.cookie || '');
-    }
-
-    let currentUser;
-    if (req.cookies.authToken) {
-      const username = jwt.verify(req.cookies.authToken, secret) as string;
-      if (username) {
-        const { rows } = await pool.query(
-          sql`SELECT * FROM users WHERE username = ${username}`
-        );
-        currentUser = rows[0];
-      }
-    }
-
-    let db;
-
-    if (!session.connection) {
-      db = await pool.connect();
-    }
-
-    return {
-      currentUser,
-      pubsub,
-      db,
-      res: session.res,
-    };
-  },
-  subscriptions: {
-    onConnect(params, ws, ctx) {
-      // pass the request object to context
-      return {
-        request: ctx.request,
-      };
-    },
-  },
+  schema: rootModule.schema,
+  context: rootModule.context,
+  subscriptions: rootModule.subscriptions,
   formatResponse: (res: any, { context }: { context: MyContext }) => {
     context.db.release();
