+++ modules/chats/index.ts
@@ -1,9 +1,12 @@
+import { GraphQLModule } from '@graphql-modules/core';
 import { gql, withFilter } from 'apollo-server-express';
 import sql from 'sql-template-strings';
+import commonModule from '../common';
+import usersModule from '../users';
 import { Message, Chat, pool } from '../../db';
 import { Resolvers } from '../../types/graphql';
 
-export const typeDefs = gql`
+const typeDefs = gql`
   type Message {
     id: ID!
     content: String!
@@ -41,7 +44,7 @@ export const typeDefs = gql`
   }
 `;
 
-export const resolvers: Resolvers = {
+const resolvers: Resolvers = {
   Message: {
     createdAt(message) {
       return new Date(message.created_at);
@@ -325,3 +328,10 @@ export const resolvers: Resolvers = {
     },
   },
 };
+
+export default new GraphQLModule({
+  name: 'chats',
+  typeDefs,
+  resolvers,
+  imports: () => [commonModule, usersModule],
+});
