+++ modules/common/index.ts
@@ -1,8 +1,12 @@
+import { GraphQLModule } from '@graphql-modules/core';
 import { gql } from 'apollo-server-express';
 import { DateTimeResolver, URLResolver } from 'graphql-scalars';
+import { pool } from '../../db';
 import { Resolvers } from '../../types/graphql';
 
-export const typeDefs = gql`
+const { PostgresPubSub } = require('graphql-postgres-subscriptions');
+
+const typeDefs = gql`
   scalar Date
   scalar URL
 
@@ -19,7 +23,34 @@ export const typeDefs = gql`
   }
 `;
 
-export const resolvers: Resolvers = {
+const resolvers: Resolvers = {
   Date: DateTimeResolver,
   URL: URLResolver,
 };
+
+const pubsub = new PostgresPubSub({
+  host: 'localhost',
+  port: process.env.DB_PORT ? parseInt(process.env.DB_PORT) : 5432,
+  user: 'testuser',
+  password: 'testpassword',
+  database: 'whatsapp',
+});
+
+export default new GraphQLModule({
+  name: 'common',
+  typeDefs,
+  resolvers,
+  async context({ res, connection }) {
+    let db;
+
+    if (!connection) {
+      db = await pool.connect();
+    }
+
+    return {
+      pubsub,
+      res,
+      db,
+    };
+  },
+});
