+++ package.json
@@ -39,6 +39,7 @@
     "typescript": "3.5.3"
   },
   "dependencies": {
+    "@graphql-modules/core": "0.7.10",
     "@safe-api/middleware": "0.0.2",
     "apollo-datasource-rest": "0.6.1",
     "apollo-server-express": "2.9.0",
