+++ tests/mutations/addChat.test.ts
@@ -1,6 +1,6 @@
 import { createTestClient } from 'apollo-server-testing';
 import { ApolloServer, PubSub, gql } from 'apollo-server-express';
-import schema from '../../schema';
+import { rootModule } from '../../index';
 import { resetDb, pool } from '../../db';
 import sql from 'sql-template-strings';
 import { MyContext } from '../../context';
@@ -12,7 +12,7 @@ describe('Mutation.addChat', () => {
     const { rows } = await pool.query(sql`SELECT * FROM users WHERE id = 2`);
     const currentUser = rows[0];
     const server = new ApolloServer({
-      schema,
+      schema: rootModule.schema,
       context: async () => ({
         pubsub: new PubSub(),
         currentUser,
@@ -69,7 +69,7 @@ describe('Mutation.addChat', () => {
     const { rows } = await pool.query(sql`SELECT * FROM users WHERE id = 1`);
     const currentUser = rows[0];
     const server = new ApolloServer({
-      schema,
+      schema: rootModule.schema,
       context: async () => ({
         pubsub: new PubSub(),
         currentUser,
