+++ tests/queries/getChat.test.ts
@@ -1,7 +1,7 @@
 import { createTestClient } from 'apollo-server-testing';
 import { ApolloServer, gql } from 'apollo-server-express';
-import schema from '../../schema';
-import { pool, resetDb } from '../../db';
+import { rootModule } from '../../index';
+import { poolm resetDb } from '../../db';
 import sql from 'sql-template-strings';
 import { MyContext } from '../../context';
 
@@ -12,7 +12,7 @@ describe('Query.chat', () => {
     const { rows } = await pool.query(sql`SELECT * FROM users WHERE id = 1`);
     const currentUser = rows[0];
     const server = new ApolloServer({
-      schema,
+      schema: rootModule.schema,
       context: async () => ({
         currentUser,
         db: await pool.connect(),
